Fix clang warnings.
This fixes all compiler warnings under clang.
The major change is safely downcasting from
Py_ssize_t to int (BSON string lengths and document
size are signed int32). This change also includes
some code cleanup in the functions touched to fix
compiler warnings.

+++ pymongo/_cmessagemodule.c
@@ -73,7 +73,7 @@ static int add_last_error(PyObject* self, buffer_t buffer,
     PyObject* one;
     char *p = strchr(ns, '.');
     /* Length of the database portion of ns. */
-    nslen = p ? (p - ns) : nslen;
+    nslen = p ? (int)(p - ns) : nslen;
 
     message_start = buffer_save_space(buffer, 4);
     if (message_start == -1) {
